requirements.txt
langchain==0.3.16
langchain-community==0.4.1
langchain-openai==0.3.4
langchain-text-splitters==1.1.1
pymupdf==1.27.1
faiss-cpu==1.10.0
sentence-transformers==5.2.3
gradio==6.6.0
python-dotenv==1.1.0
numpy<2.0.0
pydantic>=2.10.1
pydantic-core>=2.27.1